\name{CucurbitaClim}
\alias{CucurbitaClim}

\docType{data}

\title{Baseline and Future WorldClim 2.1 Climatic Data for Cucurbita Species}

\description{
This data set provides WorldClim 2.1 bioclimatic data extracted for the baseline (WorldClim 2.1 at 2.5 minutes resolution) and one future (wc2.1_2.5m_bioc_EC-Earth3-Veg_ssp245_2041-2060) climate for presence observations of Cucurbita cordata, C. digitata and C. palmata provided via the CucurbitaData data set of the \code{GapAnalysis} package. This data set is used for examples of the \code{\link[BiodiversityR]{ensemble.concave.hull}} and \code{\link[BiodiversityR]{ensemble.concave.venn}} functions.
}

\usage{
data(CucurbitaClim)
}


\references{
Fick SE and Hijmans RJ. 2017. WorldClim 2: new 1km spatial resolution climate surfaces for global land areas. International Journal of Climatology 37: 4302-4315. 

Carver et al. 2021. GapAnalysis: an R package to calculate conservation indicators using spatial information. Ecography 44: 1000-1009.
}

\examples{
data(CucurbitaClim)
}
\keyword{datasets}
