\name{warcom}
\alias{warcom}
\docType{data}
\title{Warburgia ugandensis AFLP Scores}
\description{
This data set contains scores for 185 loci for 100 individuals of the Warburgia ugandensis tree species (a medicinal tree species native to Eastern Africa). Since the data set is a subset of a larger data set that originated from a study of several Warburgia species, some of the loci did not produce bands for W. ugandensis (i.e. some loci only contain zeroes). This data set is accompanied by \code{warenv} that describes population and regional structure of the 100 individuals.
}
\usage{data(warcom)}
\format{
  A data frame with 100 observations on the following 185 variables.
  \describe{
    \item{\code{locus001}}{a numeric vector}
    \item{\code{locus002}}{a numeric vector}
    \item{\code{locus003}}{a numeric vector}
    \item{\code{locus004}}{a numeric vector}
    \item{\code{locus005}}{a numeric vector}
    \item{\code{locus006}}{a numeric vector}
    \item{\code{locus007}}{a numeric vector}
    \item{\code{locus008}}{a numeric vector}
    \item{\code{locus009}}{a numeric vector}
    \item{\code{locus010}}{a numeric vector}
    \item{\code{locus011}}{a numeric vector}
    \item{\code{locus012}}{a numeric vector}
    \item{\code{locus013}}{a numeric vector}
    \item{\code{locus014}}{a numeric vector}
    \item{\code{locus015}}{a numeric vector}
    \item{\code{locus016}}{a numeric vector}
    \item{\code{locus017}}{a numeric vector}
    \item{\code{locus018}}{a numeric vector}
    \item{\code{locus019}}{a numeric vector}
    \item{\code{locus020}}{a numeric vector}
    \item{\code{locus021}}{a numeric vector}
    \item{\code{locus022}}{a numeric vector}
    \item{\code{locus023}}{a numeric vector}
    \item{\code{locus024}}{a numeric vector}
    \item{\code{locus025}}{a numeric vector}
    \item{\code{locus026}}{a numeric vector}
    \item{\code{locus027}}{a numeric vector}
    \item{\code{locus028}}{a numeric vector}
    \item{\code{locus029}}{a numeric vector}
    \item{\code{locus030}}{a numeric vector}
    \item{\code{locus031}}{a numeric vector}
    \item{\code{locus032}}{a numeric vector}
    \item{\code{locus033}}{a numeric vector}
    \item{\code{locus034}}{a numeric vector}
    \item{\code{locus035}}{a numeric vector}
    \item{\code{locus036}}{a numeric vector}
    \item{\code{locus037}}{a numeric vector}
    \item{\code{locus038}}{a numeric vector}
    \item{\code{locus039}}{a numeric vector}
    \item{\code{locus040}}{a numeric vector}
    \item{\code{locus041}}{a numeric vector}
    \item{\code{locus042}}{a numeric vector}
    \item{\code{locus043}}{a numeric vector}
    \item{\code{locus044}}{a numeric vector}
    \item{\code{locus045}}{a numeric vector}
    \item{\code{locus046}}{a numeric vector}
    \item{\code{locus047}}{a numeric vector}
    \item{\code{locus048}}{a numeric vector}
    \item{\code{locus049}}{a numeric vector}
    \item{\code{locus050}}{a numeric vector}
    \item{\code{locus051}}{a numeric vector}
    \item{\code{locus052}}{a numeric vector}
    \item{\code{locus053}}{a numeric vector}
    \item{\code{locus054}}{a numeric vector}
    \item{\code{locus055}}{a numeric vector}
    \item{\code{locus056}}{a numeric vector}
    \item{\code{locus057}}{a numeric vector}
    \item{\code{locus058}}{a numeric vector}
    \item{\code{locus059}}{a numeric vector}
    \item{\code{locus060}}{a numeric vector}
    \item{\code{locus061}}{a numeric vector}
    \item{\code{locus062}}{a numeric vector}
    \item{\code{locus063}}{a numeric vector}
    \item{\code{locus064}}{a numeric vector}
    \item{\code{locus065}}{a numeric vector}
    \item{\code{locus066}}{a numeric vector}
    \item{\code{locus067}}{a numeric vector}
    \item{\code{locus068}}{a numeric vector}
    \item{\code{locus069}}{a numeric vector}
    \item{\code{locus070}}{a numeric vector}
    \item{\code{locus071}}{a numeric vector}
    \item{\code{locus072}}{a numeric vector}
    \item{\code{locus073}}{a numeric vector}
    \item{\code{locus074}}{a numeric vector}
    \item{\code{locus075}}{a numeric vector}
    \item{\code{locus076}}{a numeric vector}
    \item{\code{locus077}}{a numeric vector}
    \item{\code{locus078}}{a numeric vector}
    \item{\code{locus079}}{a numeric vector}
    \item{\code{locus080}}{a numeric vector}
    \item{\code{locus081}}{a numeric vector}
    \item{\code{locus082}}{a numeric vector}
    \item{\code{locus083}}{a numeric vector}
    \item{\code{locus084}}{a numeric vector}
    \item{\code{locus085}}{a numeric vector}
    \item{\code{locus086}}{a numeric vector}
    \item{\code{locus087}}{a numeric vector}
    \item{\code{locus088}}{a numeric vector}
    \item{\code{locus089}}{a numeric vector}
    \item{\code{locus090}}{a numeric vector}
    \item{\code{locus091}}{a numeric vector}
    \item{\code{locus092}}{a numeric vector}
    \item{\code{locus093}}{a numeric vector}
    \item{\code{locus094}}{a numeric vector}
    \item{\code{locus095}}{a numeric vector}
    \item{\code{locus096}}{a numeric vector}
    \item{\code{locus097}}{a numeric vector}
    \item{\code{locus098}}{a numeric vector}
    \item{\code{locus099}}{a numeric vector}
    \item{\code{locus100}}{a numeric vector}
    \item{\code{locus101}}{a numeric vector}
    \item{\code{locus102}}{a numeric vector}
    \item{\code{locus103}}{a numeric vector}
    \item{\code{locus104}}{a numeric vector}
    \item{\code{locus105}}{a numeric vector}
    \item{\code{locus106}}{a numeric vector}
    \item{\code{locus107}}{a numeric vector}
    \item{\code{locus108}}{a numeric vector}
    \item{\code{locus109}}{a numeric vector}
    \item{\code{locus110}}{a numeric vector}
    \item{\code{locus111}}{a numeric vector}
    \item{\code{locus112}}{a numeric vector}
    \item{\code{locus113}}{a numeric vector}
    \item{\code{locus114}}{a numeric vector}
    \item{\code{locus115}}{a numeric vector}
    \item{\code{locus116}}{a numeric vector}
    \item{\code{locus117}}{a numeric vector}
    \item{\code{locus118}}{a numeric vector}
    \item{\code{locus119}}{a numeric vector}
    \item{\code{locus120}}{a numeric vector}
    \item{\code{locus121}}{a numeric vector}
    \item{\code{locus122}}{a numeric vector}
    \item{\code{locus123}}{a numeric vector}
    \item{\code{locus124}}{a numeric vector}
    \item{\code{locus125}}{a numeric vector}
    \item{\code{locus126}}{a numeric vector}
    \item{\code{locus127}}{a numeric vector}
    \item{\code{locus128}}{a numeric vector}
    \item{\code{locus129}}{a numeric vector}
    \item{\code{locus130}}{a numeric vector}
    \item{\code{locus131}}{a numeric vector}
    \item{\code{locus132}}{a numeric vector}
    \item{\code{locus133}}{a numeric vector}
    \item{\code{locus134}}{a numeric vector}
    \item{\code{locus135}}{a numeric vector}
    \item{\code{locus136}}{a numeric vector}
    \item{\code{locus137}}{a numeric vector}
    \item{\code{locus138}}{a numeric vector}
    \item{\code{locus139}}{a numeric vector}
    \item{\code{locus140}}{a numeric vector}
    \item{\code{locus141}}{a numeric vector}
    \item{\code{locus142}}{a numeric vector}
    \item{\code{locus143}}{a numeric vector}
    \item{\code{locus144}}{a numeric vector}
    \item{\code{locus145}}{a numeric vector}
    \item{\code{locus146}}{a numeric vector}
    \item{\code{locus147}}{a numeric vector}
    \item{\code{locus148}}{a numeric vector}
    \item{\code{locus149}}{a numeric vector}
    \item{\code{locus150}}{a numeric vector}
    \item{\code{locus151}}{a numeric vector}
    \item{\code{locus152}}{a numeric vector}
    \item{\code{locus153}}{a numeric vector}
    \item{\code{locus154}}{a numeric vector}
    \item{\code{locus155}}{a numeric vector}
    \item{\code{locus156}}{a numeric vector}
    \item{\code{locus157}}{a numeric vector}
    \item{\code{locus158}}{a numeric vector}
    \item{\code{locus159}}{a numeric vector}
    \item{\code{locus160}}{a numeric vector}
    \item{\code{locus161}}{a numeric vector}
    \item{\code{locus162}}{a numeric vector}
    \item{\code{locus163}}{a numeric vector}
    \item{\code{locus164}}{a numeric vector}
    \item{\code{locus165}}{a numeric vector}
    \item{\code{locus166}}{a numeric vector}
    \item{\code{locus167}}{a numeric vector}
    \item{\code{locus168}}{a numeric vector}
    \item{\code{locus169}}{a numeric vector}
    \item{\code{locus170}}{a numeric vector}
    \item{\code{locus171}}{a numeric vector}
    \item{\code{locus172}}{a numeric vector}
    \item{\code{locus173}}{a numeric vector}
    \item{\code{locus174}}{a numeric vector}
    \item{\code{locus175}}{a numeric vector}
    \item{\code{locus176}}{a numeric vector}
    \item{\code{locus177}}{a numeric vector}
    \item{\code{locus178}}{a numeric vector}
    \item{\code{locus179}}{a numeric vector}
    \item{\code{locus180}}{a numeric vector}
    \item{\code{locus181}}{a numeric vector}
    \item{\code{locus182}}{a numeric vector}
    \item{\code{locus183}}{a numeric vector}
    \item{\code{locus184}}{a numeric vector}
    \item{\code{locus185}}{a numeric vector}
  }
}

\source{
Muchugi, A.N. (2007) Population genetics and taxonomy of important medicinal tree species of the genus Warburgia. PhD Thesis. Kenyatta University, Kenya.
}
\examples{
data(warcom)
}
\keyword{datasets}
